refactor : order refactor

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -30,7 +30,7 @@ public class OrderInfo {
                 String productName,
                 Long optionId,
                 String optionName,
-                BigDecimal quantity,
+                int quantity,
                 BigDecimal price,
                 String productImageUrl
         ) {
